requirements.txt
Flask==1.1.1
Flask_Classful==0.14.2
jsonpickle==1.4.1
p2pnetwork==1.1
pycryptodomex==3.11.0
requests==2.22.0
